# internal/controller/subsHandlers.go
package controller

import (
	"encoding/json"
	"fmt"
	"io"
	"net/http"
	"strings"
	"time"

	"github.com/JonnyShabli/EffectiveMobile/internal/models"
	"github.com/JonnyShabli/EffectiveMobile/internal/service"
	"github.com/JonnyShabli/EffectiveMobile/pkg/logster"
	"github.com/go-chi/chi/v5"

	_ "github.com/JonnyShabli/EffectiveMobile/docs"
)

// @title EffectiveMobile test task
// @version         1.0
// @description     Тестовое задание
// @BasePath  /api/subs

type SubsHandlerInterface interface {
	InsertSub(w http.ResponseWriter, r *http.Request)
	GetSub(w http.ResponseWriter, r *http.Request)
	UpdateSub(w http.ResponseWriter, r *http.Request)
	DeleteSub(w http.ResponseWriter, r *http.Request)
	ListSub(w http.ResponseWriter, r *http.Request)
	SumPriceByDate(w http.ResponseWriter, r *http.Request)
}

func NewSubsHandler(service service.SubsServiceInterface, log logster.Logger) SubsHandlerInterface {
	return &SubsHandler{
		Service: service,
		Log:     log,
	}
}

type SubsHandler struct {
	Service service.SubsServiceInterface
	Log     logster.Logger
}

// @Summary 		Добавление информации о подписке
// @Description 	Добавление информации о подписке
// @Tags 			Insert
// @Accept			json
// @Produce 		json
// @Param			subscription body models.SubscriptionDTO true "subscription struct"
// @Success 		200		{object}	Response
// @Failure			400		{object}	Response
// @Failure			500		{object}	Response
// @Router 			/	[post]
func (s *SubsHandler) InsertSub(w http.ResponseWriter, r *http.Request) {
	var subDTO models.SubscriptionDTO
	ctx := r.Context()
	reqBody, err := io.ReadAll(r.Body)
	if err != nil {
		s.Log.WithError(err).Infof("fail read request body")
		BadRequestResponse(w, s.Log, "fail read request body", "")
		return
	}
	err = json.Unmarshal(reqBody, &subDTO)
	if err != nil {
		err = fmt.Errorf("fail to unmarshal request body '%w'", err)
		s.Log.WithError(err).Infof("fail to unmarshal request body")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	sub := dtoToSub(&subDTO)

	sub_id, err := s.Service.InsertSub(ctx, s.Log, sub)
	if err != nil {
		if strings.Contains(err.Error(), "duplicate key value") {
			err = fmt.Errorf("duplicate key value %w", err)
			BadRequestResponse(w, s.Log, err.Error(), "")
			s.Log.WithError(err).Infof("duplicate key value")
			return
		}
		s.Log.WithError(err).Errorf("insert sub fail")
		ErrorResponse(w, s.Log, fmt.Errorf("insert sub fail %w", err).Error(), "")
		return
	}

	SuccessResponse(w, s.Log, "success", sub_id)
	s.Log.Infof("insert sub success")
}

// @Summary 		Добавление информации о подписке
// @Description 	Добавление информации о подписке
// @Tags 			Get
// @Accept			json
// @Produce 		json
// @Param			sub_id path int true "subscription ID"
// @Success 		200		{object}	Response
// @Failure			400		{object}	Response
// @Failure			500		{object}	Response
// @Router 			/{sub_id}	[get]
func (s *SubsHandler) GetSub(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	subId := chi.URLParam(r, "sub_id")
	if subId == "" {
		s.Log.Infof("fail to get sub_id")
		BadRequestResponse(w, s.Log, fmt.Errorf("fail to get sub_id").Error(), "")
		return
	}
	subs, err := s.Service.GetSub(ctx, s.Log, subId)
	if err != nil {
		if strings.Contains(err.Error(), "sql: no rows in result set") {
			err = fmt.Errorf("sql: no rows in result set %w", err)
			s.Log.WithError(err).Infof("sql: no rows in result set")
			BadRequestResponse(w, s.Log, err.Error(), "")
			return
		}
		s.Log.WithError(err).Errorf("get sub fail")
		ErrorResponse(w, s.Log, fmt.Errorf("get sub fail %w", err).Error(), "")
		return
	}
	result := subsToDTO(subs)

	SuccessResponse(w, s.Log, "success", result)
	s.Log.Infof("get sub success")
}

// @Summary 		Обновление информации о подписке
// @Description 	Обновление информации о подписке
// @Tags 			Update
// @Accept			json
// @Produce 		json
// @Param			subscription body models.SubscriptionDTO true "subscription struct"
// @Success 		200		{object}	Response
// @Failure			400		{object}	Response
// @Failure			500		{object}	Response
// @Router 			/	[put]
func (s *SubsHandler) UpdateSub(w http.ResponseWriter, r *http.Request) {
	var subDTO models.SubscriptionDTO
	ctx := r.Context()
	reqBody, err := io.ReadAll(r.Body)
	if err != nil {
		err = fmt.Errorf("fail to read request body '%w'", err)
		s.Log.WithError(err).Infof("fail to read request body")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	err = json.Unmarshal(reqBody, &subDTO)
	if err != nil {
		err = fmt.Errorf("fail to unmarshal request body '%w'", err)
		s.Log.WithError(err).Infof("fail to unmarshal request body")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	sub := dtoToSub(&subDTO)

	err = s.Service.UpdateSub(ctx, s.Log, sub)
	if err != nil {
		err = fmt.Errorf("update sub fail %w", err)
		s.Log.WithError(err).Infof("update sub fail")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	SuccessResponse(w, s.Log, "success", "")
	s.Log.Infof("update sub success")
}

// @Summary 		Удаление информации о подписке
// @Description 	Удаление информации о подписке
// @Tags 			Delete
// @Accept			json
// @Produce 		json
// @Param			sub_id path int true "subscription ID"
// @Success 		200		{object}	Response
// @Failure			400		{object}	Response
// @Failure			500		{object}	Response
// @Router 			/{sub_id}	[delete]
func (s *SubsHandler) DeleteSub(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	subId := chi.URLParam(r, "sub_id")
	if subId == "" {
		err := fmt.Errorf("fail to get sub_id")
		s.Log.WithError(err).Infof("fail to get sub_id")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	err := s.Service.DeleteSub(ctx, s.Log, subId)
	if err != nil {
		err = fmt.Errorf("delete sub fail %w", err)
		s.Log.WithError(err).Infof("delete sub fail")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	SuccessResponse(w, s.Log, "success", "")
	s.Log.Infof("delete sub success")
}

// @Summary 		Получение списка подписок
// @Description 	Получение списка подписок
// @Tags 			List
// @Accept			json
// @Produce 		json
// @Success 		200		{object}	Response
// @Failure			400		{object}	Response
// @Failure			500		{object}	Response
// @Router 			/list	[get]
func (s *SubsHandler) ListSub(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	subs, err := s.Service.ListSub(ctx, s.Log)
	if err != nil {
		if strings.Contains(err.Error(), "sql: no rows in result set") {
			err = fmt.Errorf("sql: no rows in result set %w", err)
			BadRequestResponse(w, s.Log, fmt.Errorf("sql: no rows in result set %w", err).Error(), "")
			return
		}
		s.Log.WithError(err).Errorf("list sub fail")
		ErrorResponse(w, s.Log, fmt.Errorf("list sub fail %w", err).Error(), "")
		return
	}
	result := subsToDTO(subs)

	SuccessResponse(w, s.Log, "success", result)
	s.Log.Infof("list sub success")

}

func dtoToSub(dto *models.SubscriptionDTO) *models.Subscription {
	return &models.Subscription{
		Sub_id:       dto.Sub_id,
		Price:        dto.Price,
		Service_name: dto.Service_name,
		User_id:      dto.User_id,
		Start_date:   dto.Start_date,
	}
}

// @Summary 		Подсчет суммарной стоимости всех подписок
// @Description 	За выбранный период с фильтрацией по id пользователя и названию подписки
// @Tags 			SumPriceByDate
// @Accept			json
// @Produce 		json
// @Param			SumPriceRequest body models.SumPriceRequest true "request struct"
// @Success 		200		{object}	Response
// @Failure			400		{object}	Response
// @Failure			500		{object}	Response
// @Router 			/sumPriceByDate	[get]
func (s *SubsHandler) SumPriceByDate(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	var params models.SumPriceRequest
	reqBody, err := io.ReadAll(r.Body)
	if err != nil {
		err = fmt.Errorf("fail to read request body '%w'", err)
		s.Log.WithError(err).Infof("fail to read request body")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	err = json.Unmarshal(reqBody, &params)
	if err != nil {
		err = fmt.Errorf("fail to unmarshal request body '%w'", err)
		s.Log.WithError(err).Infof("fail to unmarshal request body")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}

	start, err := time.Parse("01-2006", params.Start_date)
	if err != nil {
		err = fmt.Errorf("fail to parse start date '%w'", err)
		s.Log.WithError(err).Infof("fail to parse start date")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	end, err := time.Parse("01-2006", params.End_date)
	if err != nil {
		err = fmt.Errorf("fail to parse end date '%w'", err)
		s.Log.WithError(err).Infof("fail to parse end date")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}

	if end.Before(start) {
		s.Log.Errorf("bad interval start = %v, end = %v", start, end)
	}

	sum, err := s.Service.SumPriceByDate(ctx, s.Log, &params)
	if err != nil {
		s.Log.WithError(err).Infof("fail to sum price by date")
		BadRequestResponse(w, s.Log, err.Error(), "")
		return
	}
	SuccessResponse(w, s.Log, "sum price by date", sum)
}

func subsToDTO(subs []*models.Subscription) []*models.SubscriptionDTO {
	result := make([]*models.SubscriptionDTO, 0, len(subs))
	for _, sub := range subs {
		result = append(result, &models.SubscriptionDTO{
			Sub_id:       sub.Sub_id,
			Price:        sub.Price,
			Service_name: sub.Service_name,
			User_id:      sub.User_id,
			Start_date:   sub.Start_date,
		})
	}

	return result
}